✨ no extension methods
