+++ zio-parser/shared/src/main/scala-2.x/zio/parser/SyntaxCompanionVersionSpecific.scala
@@ -1,0 +1,9 @@
+package zio.parser
+
+trait SyntaxCompanionVersionSpecific {
+  def oneOf[A, I, O](
+    syntax: Syntax[String, I, O, _ <: A],
+    syntaxes: Syntax[String, I, O, _ <: A]*,
+  ): Syntax[String, I, O, A] = 
+    syntaxes.map(_.widen[A]).foldLeft(syntax.widen[A])(_ | _)
+}
